requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
coverage==7.5.4
celery==5.4.0
redis==5.0.7
gevent==24.2.1
requests==2.32.3
python-telegram-bot==21.3
pytz==2024.1
drf-yasg==1.21.7
